Enhancement: Extract MinimumCountResolutionStrategy
